+++ src/api/updatePhoneIndex.ts
@@ -1,3 +1,4 @@
+import { accountStore } from '../stores/accountStore'
 import { getAuthStore } from '../stores/authStore'
 import { intl, intlDebug } from '../stores/intl'
 import { Nav } from '../stores/Nav'
@@ -24,12 +25,21 @@ export const updatePhoneIndex = async (
     console.error('updatePhoneIndex.setExtensionProperties: extProps undefined')
     return null
   }
+
+  // upsert account info with properties phoneappli.enable
+  const as = getAuthStore()
+  if (p.id === as.getCurrentAccount().id) {
+    as.userExtensionProperties = extProps
+    const d = await as.getCurrentDataAsync()
+    d.phoneappliEnabled = extProps.phoneappli
+    accountStore.updateAccountData(d)
+  }
+
   const phone = extProps.phones[phoneIndex - 1]
   const phoneTypeCorrect = phone.type === 'Web Phone'
   const { pbxTenant, pbxUsername } = p
   const expectedPhoneId = `${pbxTenant}_${pbxUsername}_phone${phoneIndex}_webphone`
   const phoneIdCorrect = phone.id === expectedPhoneId
-  //
   const setExtensionProperties = async () => {
     if (!api.client) {
       console.error(
@@ -46,11 +56,8 @@ export const updatePhoneIndex = async (
         [`p${phoneIndex}_ptype`]: phone.type,
       },
     })
-    if (p === getAuthStore().getCurrentAccount()) {
-      getAuthStore().userExtensionProperties = extProps
-    }
   }
-  //
+
   if (phoneTypeCorrect && phoneIdCorrect) {
     // do nothing
   } else if (phoneTypeCorrect && !phoneIdCorrect) {