+++ src/components/App.tsx
@@ -115,14 +115,15 @@ const initApp = async () => {
     s.resetFailureState()
     cs.onCallKeepAction()
     pnToken.syncForAllAccounts()
+    if (checkHasCallOrWakeFromPN() || (await s.handleUrlParams())) {
+      return
+    }
     // with ios when wakekup app, currentState will be 'unknown' first then 'active'
     // https://github.com/facebook/react-native-website/issues/273
     if (Platform.OS !== 'ios') {
       return
     }
-    if (!checkHasCallOrWakeFromPN() && !(await s.handleUrlParams())) {
-      await autoLogin()
-    }
+    await autoLogin()
   })
 
   NetInfo.addEventListener(({ isConnected }) => {
@@ -176,7 +177,6 @@ const initApp = async () => {
   }, 17)
   const clearReaction = reaction(() => s.signedInId, onAuthUpdate)
   void clearReaction
-
   if (await s.handleUrlParams()) {
     console.log('App navigated by url params')
     // already navigated