+++ src/components/HeaderDropdown.tsx
@@ -49,7 +49,7 @@ const css = StyleSheet.create({
     top: 0,
     bottom: 0,
     right: 0,
-    width: 50,
+    width: 40,
   },
 })