+++ src/pages/PageCustomPageView.tsx
@@ -2,10 +2,10 @@ import { observer } from 'mobx-react'
 import { Component } from 'react'
 import { StyleSheet } from 'react-native'
 
+import { isCustomPageUrlBuilt } from '../api/customPage'
 import {
-  isCustomPageUrlBuilt,
-  rebuildCustomPageUrl,
   rebuildCustomPageUrlNonce,
+  rebuildCustomPageUrlPbxToken,
 } from '../api/pbx'
 import { PbxCustomPage } from '../brekekejs'
 import { CustomPageWebView } from '../components/CustomPageWebView'
@@ -54,7 +54,7 @@ export class PageCustomPageView extends Component<{ id: string }> {
     if (!cp) {
       return
     }
-    const url = await rebuildCustomPageUrl(cp.url)
+    const url = await rebuildCustomPageUrlPbxToken(cp.url)
     as.updateCustomPage({ ...cp, url })
   }
 
@@ -102,13 +102,16 @@ export class PageCustomPageView extends Component<{ id: string }> {
       s.name == 'PageCustomPage' &&
       RnStacker.stacks.length == 1
 
-    const loading = this.state.jsLoading || this.state.webviewLoading
+    // onLoadEnd not fire with website load image from url camera
+    // so, should be check loading like bellow
+    const loaded = !this.state.jsLoading || !this.state.webviewLoading
+
     const title = cp?.title
       ? cp.title
-      : loading
+      : !loaded
       ? intl`Loading...`
       : intl`PBX user settings`
-    const description = loading
+    const description = !loaded
       ? intl`Loading...`
       : this.state.webviewError
       ? // TODO
@@ -140,6 +143,7 @@ export class PageCustomPageView extends Component<{ id: string }> {
               this.setState({
                 webviewLoading: false,
                 webviewError:
+                  e &&
                   'code' in e.nativeEvent &&
                   typeof e.nativeEvent.code === 'number',
               })