+++ src/utils/deeplink-parse.ts
@@ -54,6 +54,7 @@ export interface UrlParams {
   port: string
   phone_idx: string
   _wn: string
+  number: string | undefined
   action: string
   callTo: string
 }